==> notebooks/1-data_preprocessing.ipynb <==
import numpy as np
import pandas as pd
import os

df_pesticides = pd.read_csv('../dataset/pesticides.csv')
df_rainfall = pd.read_csv('../dataset/rainfall.csv')
df_temperature = pd.read_csv('../dataset/temperature.csv')
df_yield = pd.read_csv('../dataset/yield.csv')

print('Pesticides: \n')
display(df_pesticides.head())
print('Precipitation/Rainfall: \n')
display(df_rainfall.head())
print('Temperature: \n')
display(df_temperature.head())
print('Yield: \n')
display(df_yield.head())

# take only the necessary fields
df_yield = df_yield[['Area', 'Item', 'Year', 'Value']]
df_pesticides = df_pesticides[['Area', 'Year', 'Value']]
# change the name of columns (if necessary) in order to merge them later
df_temperature.rename(columns = {'year' : 'Year', 'country' : 'Area'}, inplace = True)

df_temperature = df_temperature[df_temperature.Year >= 1961]

# cleaning and merging data
merged_data = pd.merge(df_yield, df_temperature, on =['Year', 'Area'])
merged_data = pd.merge(merged_data, df_pesticides, on = ['Year', 'Area'])

merged_data.head()

# Since there is a column by the name of Value in df_yield and df_pesticide. then we have the x and y value
merged_data.rename(columns={'Value_x': 'Yield', 'Value_y': 'Pesticides'}, inplace=True)

# remove the space in column name
df_rainfall.rename(columns = {' Area': 'Area',}, inplace = True)

merged_data = pd.merge(merged_data, df_rainfall, on=['Year', "Area"])
merged_data.head()

# change the rainfall data column to short and meaningful 
merged_data.rename(columns = {'average_rain_fall_mm_per_year':'avg_precipitation'}, inplace = True)

# saving the merged data
# os.path.join(path, file_name)
path_to_save = os.path.join('../dataset', 'merged_data.csv')
merged_data.to_csv(path_to_save, index=False)
