==> notebooks/4-tunning_hyperparameters.ipynb <==
### Tunning Hyperparameters in previous tried algorithms

def perform_grid_search(estimator, param_grid, model_name, X_train, y_train, X_test, y_test):
    grid_search = GridSearchCV(estimator=estimator, param_grid=param_grid, cv=5, scoring='neg_mean_absolute_error', n_jobs=-1)
    grid_search.fit(X_train, y_train)
    
    best_estimator = grid_search.best_estimator_
    predictions = best_estimator.predict(X_te
    
    best_mae = -grid_search.best_score_  # neg_mean_absolute_error is returned negative by GridSearchCV
    mse = mean_squared_error(y_test, predictions)
    rmse = sqrt(mse)
    r2 = r2_score(y_test, predictions)
    
    print(f"{model_name} Best Parameters: {grid_search.best_params_}")
    print(f"{model_name} Best MAE: {best_mae}")
    print(f"{model_name} Best MSE: {mse}")
    print(f"{model_name} Best RMSE: {rmse}")
    print(f"{model_name} Best R-squared: {r2}")


# Random Forest Grid Search
rf_params = {
    'n_estimators': [100, 200, 300],
    'max_depth': [None, 10, 20],
    'min_samples_split': [2, 5, 10]
}
perform_grid_search(RandomForestRegressor(random_state=42), rf_params, 'Random Forest', X_train, y_train, X_test, y_test)

# Gradient Boosting Grid Search
gb_params = {
    'n_estimators': [100, 200, 300],
    'learning_rate': [0.01, 0.1, 0.2],
    'max_depth': [3, 5, 10]
}
perform_grid_search(GradientBoostingRegressor(random_state=42), gb_params, 'Gradient Boosting', X_train, y_train, X_test, y_test)

# Decision Tree Grid Search
dt_params = {
    'max_depth': [None, 10, 20, 30],
    'min_samples_split': [2, 5, 10],
    'min_samples_leaf': [1, 2, 4]
}
perform_grid_search(DecisionTreeRegressor(random_state=42), dt_params, 'Decision Tree', X_train, y_train, X_test, y_test)

# K-Nearest Neighbors Grid Search
knn_params = {
    'n_neighbors': [3, 5, 7],
    'weights': ['uniform', 'distance'],
    'algorithm': ['auto', 'ball_tree', 'kd_tree', 'brute']
}
perform_grid_search(KNeighborsRegressor(), knn_params, 'K-Nearest Neighbors', X_train, y_train, X_test, y_test)

nn_params = {
    'hidden_layer_sizes': [(50,), (100,), (50,50), (100,100)],
    'activation': ['identity', 'logistic', 'tanh', 'relu'],
    'solver': ['lbfgs', 'sgd', 'adam'],
    'alpha': [0.0001, 0.001, 0.01, 0.1],
    'learning_rate': ['constant', 'invscaling', 'adaptive'],
    'max_iter': [200, 500, 1000]
}

# Perform grid search for MLPRegressor
perform_grid_search(MLPRegressor(random_state=42), nn_params, 'Neural Network', X_train, y_train, X_test, y_test)

I stopped the training because it was very resource demanding and I have only standard laptop
